requirements.txt
anyio==3.6.2
autopep8==2.0.1
bidict==0.22.1
click==8.1.3
colorama==0.4.6
fastapi==0.91.0
future==0.18.3
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycodestyle==2.10.0
pydantic==1.10.4
pynput==1.7.6
pyserial==3.5
simple-websocket==0.9.0
uvicorn==0.20.0
websockets==10.4
Werkzeug==2.2.2
